Collapse /users/:user_id and /users/:user_id/crates
We don't need a distinct user profile page at the moment.
