+++ app/router.js
@@ -27,9 +27,7 @@ Router.map(function() {
         this.route('crates');
         this.route('following');
     });
-    this.route('user', { path: '/users/:user_id' }, function() {
-        this.route('crates');
-    });
+    this.route('user', { path: '/users/:user_id' });
     this.route('install');
     this.route('search');
     this.route('dashboard');
