+++ app/templates/user.hbs
@@ -1,16 +1,68 @@
 <div id='crates-heading'>
     {{#user-link user=model}}
-        {{user-avatar user=model size='medium'}}
+        {{user-avatar user=model.user size='medium'}}
     {{/user-link}}
     <h1>
-        {{#link-to 'user' model}}
-            {{ model.login }}
+        {{#link-to 'user' model.user.login}}
+            {{ model.user.login }}
         {{/link-to}}
     </h1>
 </div>
 
 <div id='user-profile'>
     <div class='info'>
-      {{outlet}}
+        {{! TODO: reduce duplication with templates/crates.hbs }}
+
+        <div id='results'>
+            <div class='nav'>
+                <span class='amt small'>
+                    Displaying
+                    <span class='cur'>{{ currentPageStart }}-{{ currentPageEnd }}</span>
+                    of <span class='total'>{{ totalItems }}</span> total results
+                </span>
+            </div>
+
+            <div class='sort'>
+                <span class='small'>Sort by</span>
+                {{#rl-dropdown-container class="dropdown-container"}}
+                    {{#rl-dropdown-toggle tagName="a" class="dropdown"}}
+                        <img class="sort" src="/assets/sort.png"/>
+                        {{ currentSortBy }}
+                        <span class='arrow'></span>
+                    {{/rl-dropdown-toggle}}
+
+                    {{#rl-dropdown tagName="ul" class="dropdown" closeOnChildClick="a:link"}}
+                        <li>
+                            {{#link-to (query-params sort="alpha")}}
+                                Alphabetical
+                            {{/link-to}}
+                        </li>
+                        <li>
+                            {{#link-to (query-params sort="downloads")}}
+                                Downloads
+                            {{/link-to}}
+                        </li>
+                    {{/rl-dropdown}}
+                {{/rl-dropdown-container}}
+            </div>
+        </div>
+
+        <div id='crates' class='white-rows'>
+            {{#each model.crates as |crate|}}
+                {{crate-row crate=crate}}
+            {{/each}}
+        </div>
+
+        <div class='pagination'>
+            {{#link-to (query-params page=prevPage) class="prev" rel="prev" title="previous page"}}
+                <img class="left-pag" src="/assets/left-pag.png"/>
+            {{/link-to}}
+            {{#each pages as |page|}}
+                {{#link-to (query-params page=page)}}{{ page }}{{/link-to}}
+                {{/each}}
+                {{#link-to (query-params page=nextPage) class="next" rel="next" title="next page"}}
+                    <img class="right-pag" src="/assets/right-pag.png"/>
+                {{/link-to}}
+        </div>
     </div>
 </div>
