--- app/templates/user/index.hbs
@@ -1,7 +1,0 @@
-<p>
-    Generic profile information here.
-</p>
-
-{{#link-to 'user.crates' model.login}}
-    See all their crates
-{{/link-to}}
